+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`172` add new `FlowCytometry` instruction and corresponding protocol method
 * :feature:`174` use more sensible default z positions for pre_buffer and blowout in `LiquidHandleMethod`
 * :bug:`174` fix broken PreMixBlowoutTransfer that used outdated logic
 * :feature:`170` protect liquid_handle-related utils until they can be made more general-purpose